Core functionality implemented
fetching from the api and showing the question with answers works.

when clicking the wrong answer the border turns red, and the border of the correct answer turns green with a check mark, then a next button appears to go to the next question, when the last question is appeared the button that appears after the question is answered is 'finish' which shows the quiz results page.